requirements.txt
fastapi[standard]
pymongo
python-dotenv
python-multipart
uvicorn
bcrypt